refactor: extract vaadin-split-layout logic to reusable mixin
